fix(cli): restore CI green — split session-start TU + test link stubs
Two regressions slipped in while `testing` was unprotected:

1. cli_main.c grew to 2151 lines, past the 2000-line hard limit
   (lint + build-integrity). Extract the SessionStart-hook entry
   (handle_session_start + the remote proactive-recall fallback and
   its ss_sbuf/ss_render_section helpers) into a new cli_session_start.c
   (cli_main.c -> 1903). read_stdin / client_hook_payload_session_id are
   un-static'd and shared via the new cli_session_start.h (still used by
   the pre/post-tool hook handlers left in cli_main.c). Wired into both
   build systems (Makefile CLI_SRCS + CMakeLists CLI_SRCS).

2. unit-test-cli-mcp-serve failed to link: cli_mcp_serve.c's
   server_request gained a remote-routing branch referencing
   cli_rpc_has_remote_endpoint / cli_rpc_client_endpoint /
   cli_rpc_client_bearer / cli_v1_route_for_method / cli_http_request
   (plus the reverse-channel start/stop). Add standalone stubs in
   test_cli_mcp_serve.c (the test drives the local-socket path, so the
   remote branch is never taken).

Also update test_cli_http_transport: delegate now resolves to
POST /v1/delegate/run (Phase 4), so it can no longer stand in as the
"unmapped method" example — assert the real mapping and use a fake
method for the NULL case.

Local: unit-tests, build-integrity (49/0), check-cli-v1-routes,
check-v1-method-coverage, clang-format-19, and the aimee CLI build all green.
